User type changed

+++ Task01.py
@@ -20,7 +20,7 @@ new_entry = pd.DataFrame({'Sample_ID':['L0071'],
             })
 print(new_entry)
 
-user_position = int(input("Enter position to insert: "))
+user_position = float(input("Enter position to insert: "))
 if user_position < 0 or user_position > (len(cinnamon_data)+1): #not below 0 index and not leaving cells empty at the end of dataframe
         raise ValueError("Invalid insertion index!")
 elif 0 < user_position < len(cinnamon_data):
